Added projections to eventstore that combine all events (for simple number based progress tracking). Added code to make dapper inserts and updates indempotent.

+++ Eventualize.TestDatabase/Tasks.sql
@@ -4,7 +4,7 @@
 	[Id] uniqueidentifier not null,
 	[Title] varchar(50) not null,
 	[Description] varchar(200) null,
-	[Version] int not null,
+	[Version] bigint not null,
 	[LastEventDate] datetime2 not null, 
     [LastEventNumber] BIGINT NOT NULL, 
     [LastModifierId] VARCHAR(50) NOT NULL
